backend/requirements.txt
# telling Python what packages our API needs

fastapi==0.115.0
uvicorn[standard]==0.30.6

sqlalchemy==2.0.34
psycopg2-binary==2.9.9
alembic==1.13.2

pydantic==2.9.2
pydantic-settings==2.5.2

python-multipart==0.0.9
python-jose==3.3.0
requests==2.32.3
google-generativeai

google-genai
redis
rq
google-cloud-storage
openai-whisper
moviepy
torch
